gcc_comp_m2_v3/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
pytest==7.4.4

# Logging and utilities
structlog==24.1.0

# Vault integration (primary service)
hvac==2.1.0

# Encryption
cryptography==41.0.7

# TLS verification
certifi==2023.11.17

# Secondary services (credentials managed by Vault)
openai==1.10.0
pinecone-client==3.0.0

# PostgreSQL support
psycopg2-binary==2.9.9

# Kubernetes client (for ServiceAccount auth)
kubernetes==28.1.0

# Additional utilities
requests==2.31.0
